🧬 Migrate the logic about processing response value into enum object *ResponseStrategy* because they're relative with it.
